+++ src/Symfony/Bundle/FrameworkBundle/DependencyInjection/Compiler/UnusedTagsPass.php
@@ -23,9 +23,9 @@ class UnusedTagsPass implements CompilerPassInterface
 {
     private const KNOWN_TAGS = [
         'annotations.cached_reader',
-        'assets.package',
         'asset_mapper.compiler',
         'asset_mapper.importmap.resolver',
+        'assets.package',
         'auto_alias',
         'cache.pool',
         'cache.pool.clearer',
@@ -33,7 +33,6 @@ class UnusedTagsPass implements CompilerPassInterface
         'chatter.transport_factory',
         'config_cache.resource_checker',
         'console.command',
-        'container.do_not_inline',
         'container.env_var_loader',
         'container.env_var_processor',
         'container.excluded',